+++ Angular-Frontend/src/app/component/Base_Components/notification/notification.component.ts
@@ -2,7 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { MycookiesService } from '../../Admin/mycookies.service';
 import { Router } from '@angular/router';
-import { NgFlashMessageService } from 'ng-flash-messages';
+import { MatSnackBar, MatDialog, MatSnackBarConfig } from '@angular/material';
+import { ConfirmationDialogComponent } from '../../Auth/confirmation-dialog/confirmation-dialog.component';
 
 interface notification {  //decalare interface class for load notification attributes.
   _id: String;
@@ -38,7 +39,8 @@ export class NotificationComponent implements OnInit {
     private http: HttpClient,
     private router: Router,
     private cookies: MycookiesService,
-    private ngFlashMessage: NgFlashMessageService,
+    public snackBar: MatSnackBar,
+    private dialog: MatDialog,
   ) { }
 
   ngOnInit() {
@@ -68,26 +70,36 @@ export class NotificationComponent implements OnInit {
     var urlDelete = "http://localhost:3000/notification/notAttachment"; //notification attachment delete url
 
     //if there is a file in attachment call atachment file delteing request
-    if(mybtnFile){
-      this.http.delete(urlDelete + '/' + mybtnFile).subscribe(res => {
-        console.log(res);
-      }, (err) => {
-        console.log(err)
-      });
-    }
-    //call content delete request
-    this.http.delete(url + '/' + mybtnId).subscribe(res => {  //send delete the notification request to the server
-      this.ngFlashMessage.showFlashMessage({
-        messages: ["Successfully Added ..!"],
-        dismissible: true,
-        timeout: 2000,
-        type: 'success',
-      });
-    }, (err) => {
-      console.log(err);
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: {
+        message: 'Are you sure want to Disapprove?',
+        buttonText: {
+          ok: 'Yes',
+          cancel: 'No'
+        }
+      }
     });
+    dialogRef.afterClosed().subscribe((confirmed: boolean) => {
+      if (confirmed) {
+        if (mybtnFile) {
+          this.http.delete(urlDelete + '/' + mybtnFile).subscribe(res => {
+            console.log(res);
+          }, (err) => {
+            console.log(err)
+          });
+        }
+        //call content delete request
+        this.http.delete(url + '/' + mybtnId).subscribe(res => {  //send delete the notification request to the server
+          let config = new MatSnackBarConfig();
+          config.duration = true ? 2000 : 0;
+          this.snackBar.open("Successfully Disapproved..! ", true ? "Done" : undefined, config);
+        }, (err) => {
+          console.log(err);
+        });
 
-    window.location.reload();     // reload the page
+        window.location.reload();     // reload the page
+      }
+    })
   }
 
   approve(event, notice_id) {     //approve button action
@@ -95,14 +107,28 @@ export class NotificationComponent implements OnInit {
     console.log(mybtnId);
 
     var url = "http://localhost:3000/notification/approve";
-
-    this.http.get(url + '/' + mybtnId).subscribe(res => {  //send add a notification request to the server
-      console.log(res);
-      alert("Successfully Approved..!");
-      window.location.reload();   //realod window
-    }, (err) => {
-      console.log(err);
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: {
+        message: 'Are you sure want to Approve?',
+        buttonText: {
+          ok: 'Yes',
+          cancel: 'No'
+        }
+      }
     });
+    dialogRef.afterClosed().subscribe((confirmed: boolean) => {
+      if (confirmed) {
+        this.http.get(url + '/' + mybtnId).subscribe(res => {  //send add a notification request to the server
+          console.log(res);
+          let config = new MatSnackBarConfig();
+          config.duration = true ? 2000 : 0;
+          this.snackBar.open("Successfully Approved..! ", true ? "Done" : undefined, config);
+          window.location.reload();   //realod window
+        }, (err) => {
+          console.log(err);
+        });
+      }
+    })
   }
 }
